Update README for Laravel Australia Post Package
This commit significantly expands the README for the Laravel Australia Post Package. It includes sections on package description, requirements, installation, and usage. In addition, detailed examples of how to use the package have been provided. This comprehensive update aims to improve user+support documentation, making it easier for others to integrate and use the Australia Post API in their Laravel applications.